MythCenter-wide: Theme the 'Metadata Options' screen.
This adds a version of the Metadata Options screen that is more consistent with
the other schedule editor screens, it adds the common elements from the other
screens like program title and description and also adds the coverart, fanart
and banner images so you can see what images have been found.

This also shows one of the first uses of the <group> widget. All the common
widgets on the schedule editor screens are defined once in a <group> which is
then added to each editor screen. Changing the layout in the group will then
change the layout in all the editor screens that use the group.