MythCenter: Theme the 'Metadata Options' screen.
This adds a version of the Metadata Options screen that is more consistent with
the other schedule editor screens, it adds the common elements from the other
screens like program title and description and also adds the coverart, fanart
and banner images so you can see what images have been found.

This also tidies the layout of some of the schedule editor screens.