Remove the "Confirm exit" setting.
The "Confirm exit" setting is another setting that changes the behavior
of an action based on a setting, and, therefore, should actually be
handled through binding the exit key to the appropriate action.

This patch adds a new action Main Menu/EXITPROMPT which displays a
prompt before exiting and uses the existing Main Menu/EXIT action to
exit without prompting the user.  This allows users to choose whether
mythfrontend should prompt them before exit and even allows a user to
bind different keys to each action so they can exit with a prompt in
some cases (such as when using Escape key, often bound to the Stop
button on remotes) and exit without a prompt in other cases (such as
when using the key bound to the Power button on their remote or a
keyboard shortcut that's not mapped on their remote).

Key bindings are updated automatically based on the currently-specified
"Confirm exit" setting such that the key mapped to EXIT is re-mapped to
EXITPROMPT is "Confirm exit" is enabled.  The default behavior for new
MythTV systems is unchanged--EXITPROMPT will be mapped to the Escape key
since the "Confirm Exit" setting defaulted to enabled.

Thanks to [R] on IRC for the idea, and to skd5aner and stuartm on IRC
for the final push to finish and the inspiration to figure out that last
10% of the patch.