Extend the Services Lookup API.
Adds most of the rest of useful metadata to a lookup return, including all relevant artwork URLs.  A returned item now comes with links to the thumbnail and raw image for each available artwork type.

Should now even be possible to implement a artwork downloader/chooser in a third party app if one were so inclined.  Yes, I mean you.

My personal approach to that would be to present the user with the thumbnails of each type, pick the ones they want with radio buttons, then use the MythDownloadManager protocol command to have the backend download the artwork with the correct filename (inetref_artworktype.ext for movies, Title Season SeasonNumber_artworktype.ext for TV) and then set up a row in the recordedartwork table.