Support recording artwork in the OSD.
Themers can now use fanart, banner, coverart, and screenshot imagetypes in the OSD and expect them to behave identically in both recordings and videos (as opposed to only working for videos).

This looks like a LOT larger change than it actually is.  In reality, it's a number of changes to the media plugin registration, a header move, and just a couple dozen lines to add support for the imagery to the OSD.