Alter mythcommflag --video behavior
In --video mode, mythcommflag currently requires a myth:// URI, absolute
file path or path relative from the current directory.  This adds a
fallthrough behavior where if a local path cannot be found, it attempts
to access it as a myth:// URI at the local hostname.  This is half of
the fix for #9435, the second half will require ProgramInfo to properly
store seekdata for such files.

Refs #9435