Idiot proof logging thread dtors.
Basically this is the same thing I applied to all the threads that were ported from pthread to QThread.

It's an error if the threads are still running when we enter the thread dtor, this makes sure the threads are stopped if they we somehow missed waiting on them to stop before calling the dtor.

I experienced the error condition only with LoggerThread, but the consequences of deleting a thread's stack while it is still running are dire and the fix quite simple and safe so I applied it to both threads.