MythNetvision:  Stream videos in the Internal player.
When a video's XML or RSS provides a download link, make the default selection action to stream it in the Internal player, now that the streaming ringbuffer is working pretty reliably.

This makes all Video RSS subscriptions, Twit.tv, Ted Talks, and the other download-providing grabbers work really nicely and much more integrated-feeling.

So, for reference, we now have three ways of tackling this content:

1) We have access to the video file, and can stream directly to our player.
2) We don't have a download, but the flash player has an API.  Play fullscreen and support remote controls.
3) Provider is unenlightened.  Spawn a web browser and play there.