Split mythlogging.h into parts to reduce recompiles
Causes an API version bump.  Make distclean likely required.

As everybody and its dog uses mythlogging.h, any small changes to internals
would precipitate a recompile of nearly everything.  This is not fun.  So...

mythlogging.h remains the header presenting the logging API to the rest of the
              system
loggingdefs.h contains the definition of LogLevel_t enum (likely to be
              combined with verbosedefs.h)
logging.h     contains the internals of the logging system (classes, etc)
logging.cpp   contains the code for the logging system

I had to rename mythlogging.cpp -> logging.cpp or moc wouldn't create the
moc_logging.cpp file (or moc_mythlogging.cpp) as the class definitions weren't
in mythlogging.h anymore.