API: Add GetImageFile, allow any artwork to be resized.
This adds a special GetImageFile method, and makes the existing means of getting recording and video artwork use it.

You can now request a fanart, banner, coverart, or screenshot for a given recording or video and specify a width, height, or both (better to specify one to preserve aspect ratio).  A new image is created in the appropriate storage group, where it remains cached (not unlike screenshots for mythweb).