Rework LIRC::GetCodes() to simplify buffer usage
Fixes #9246

This started as a fix for a valgrind-found branch based on uninitialized data.
As I was in there trying to make it not be branching based on the last
character after expanding the buffer, I simplified the algorithm as well.

Now, we resize the buffer to hold 128 new bytes just before doing the read,
and always read up to 128 bytes.  Instead of resizing after, I changed it to
a truncate, and use takeLast() rather than back() and pop_back() on the
returned list.