Add testing methods to metadata classes.
Add a means of testing the movie and television grabbers for functionality, and tie them together in the metadata factory.  This way we can do a baseline test of grabbers before a batch execution on a system that might be misconfigured.

Not actually triggered yet, but when the grabbers support the -t argument it can be triggered in mythmetadatalookup before a run so that a broken configuration can avoid a multithousand line error log and instead display a single plain english log about what is broken.