OK, third time's the charm.
Logging for console at LOG_ERR is only when progress is meant to be on the
console.  Not when running in daemon mode (or when the top application is)...

A bit of fiddly changes to the logic, but now it's seeming to be right as best
I can tell (currently running my backend in --daemon mode to catch any
outliers as well).