Fix exit after changing themes.
In caea1c7d71 , the ThemeChooser's calls to the Reload Theme jump point
were modified to include the pop = false argument so that the "old"
theme's main menu screen wasn't shown prior to loading the new theme.
Unfortunately, this left 2 extra screens on the stack, so we would never
again attempt to exit because it looked like we weren't at the top of
the main menu.

I am working on another approach that will provide behavior closer to
that intended in caea1c7d71, but it's grown large and complicated enough
that it will take a while to complete, so I'm committing this change
while I finish the other approach.

Fixes #9936.