Allow users to reset all keys to defaults.
This patch adds a menu item in MythControls (Utilities/Setup|Edit Keys)
to "Reset All Keys to Defaults".  The patch also adds a jump point,
"Reset All Keys," which does the same.

Note that the bindings for plugins will not be reset until the next time
mythfrontend is started, but they will continue to work as they were
previously mapped.  They will also not show in MythControls (nor the
database) until a frontend restart--when the plugins are re-init'ed.  I
didn't think this warranted a change to the plugin API nor unloading and
reloading all plugins just for the once-in-a-blue-moon "Reset All Keys".

The jump point was primarily to allow access to all the static functions
used in mythfrontend's main.cpp for key bindings initialization and jump
point callbacks from the MythControls code, but could actually be quite
useful if someone breaks "critical" key bindings like those which allow
navigating the menu to get to MythControls.  I did not define a default
mapping for the jump point, but with MythWeb's key bindings/jump point
mapping editor, users should be able to recover even if they didn't
assign a key before breaking bindings.

While I was in there, I renamed the ResetKeys() functions I had added in
anticipation of this patch to ReloadKeys() to better describe their
purpose.