Prevent UPnP detection from using backends of a different version
This implements the new ConnectionInfo values into the DatabaseParams
struct, and throws an error if the remote backend is found to have a
different protocol or database version.

Closes #9174

While this does not address the secondary issue that OSX does not
support detection of an interactive console, it does prevent a UPnP
connected client from ever altering the database.