Revert "Make mythgallery's thumb generator not auto-detect"
This reverts commit d1ab62dc3dee829e75dd649b36adc875ca06643a.

Ref #9523

Seems that this well-intentioned patch does in fact break the thumbnails for
directories completely.  As this is the case, I'm removing it.  We do plan to
completely rework mythgallery to be using the mythui throughout, and perhaps
at that time, we can look at making it more efficient.  In the meantime, this
is borked, so it's coming out.