Fix deleting expired recordings in the status screen.
There was a bug introduced by https://github.com/MythTV/mythtv/commit/48fb2493
which was preventing recordings from being deleted or their recording group
being changed.

This changes the expire button list back to reloading after something is
changed, each item in the list needs to be updated to reflect the new stats
anyway so there is little advantage to not reloading the list completely. To
speed things up we only grab the expire list once from the BE. We also restore
the position after refreshing the button list which was the original complaint
being fixed by 48fb2493.

Refs #9286, #8787.