Fixes an issue where some RecordingInfo() constructions on the backend were not fully populated.
To avoid an event loop in ProgramInfo::LoadFromScheduler() we simply ignored this data when called from within a backend.
Thankfully an error message was printed out which is what allowed me to find the problem to fix it.

In theory this data was already available to us in the master backend through the scheduler, but we did not have any simple means of communicating with the scheduler. This adds a MythScheduler interface and adds an accessor for it to MythCoreContext. There are only two methods made available in it, GetAllPending() and GetRecording(), but this could also be extended to get rid of thunking classes like ScheduledRecording (which sends a one-way signal to the scheduler).

This also gets rid of the global sched variable. The few places needed to be a global now use the MythScheduler interface instead.