Fix channel browsing mode when video not playing.
MythPlayer pausing has changed (for the better), but the channel browsing code was assuming that if MythPlayer::IsPaused() returned true it meant that compositing (the OSD) was disabled. Now the OSD is always active and IsPaused() just means the video itself is not playing. Anyway, we just need to remove the old sanity check code and browse mode works properly again.