Fix MythGallery's use of removable devices
MythGallery has a menu option 'Show Devices' to show pictures on removable
media. However, there are some bugs in the current code that stop this from
working and also prevent the display of pictures on media inserted if
MediaMonitor 'Jump to Plugin' is enabled.

This also effectively enables the media monitor for Windows builds.

Closes #9522

Signed-off-by: Gavin Hurlbut <ghurlbut@mythtv.org>