Use the ${ln_s} variable so config.h and config.mak are always updated when using MinGW.
Since MinGW (on windows at least) does not support symlinking natively it makes a copy instead of a link, new compiles are not updated by 'ln -s' as the file already exists, by forcing it it will update the file.
As ${ln_s} already defines to use 'ln -sf' I made the change to use this variable.