Use the ${ln_s} variable so mythconfig.mak is always updated when using MinGW.
Since MinGW (on windows at least) does not support symlinking natively it makes a copy instead of a link, new compiles are not updated by 'ln -s' as the file already exists, by forcing it it will update the file.
Since ${ln_s} is already defined in the core configure I also defined it here and set it to 'ln -sf'.