MythGame: Add a intid to Game Metadata DB.
This adds an auto-increment primary key to the game metadata DB, equivalent to the one in video metadata.

Also adds a method to get all the Rom Info from the DB into a QList, which will be of use in writing a better game scanner (and for dissecting the DB queries and UI in the hopefully near future).