Fix HTTP Proxy usage.
When we setup the http proxy, we were using as a type HttpProxy, when it should
have been HttpCachingProxy (as we are using it *only* for HTTP requests).  This
allowed when we did a listen that it would try to proxy the incoming connections
as far as I can tell.

Also, to be absolutely sure, when we use a MythServer or the HTTPServer, we
now set the connection's proxy to NoProxy before calling the listen.  This
keeps it from even trying.

Fixes #9970