Support reading of ID3v2 or ID3v1 tags in FLAC
If an ID3 tag exists then we will read and write to that instead of the
Vorbis comment tag. This includes support for albumart in ID3v2 tags
found in FLAC files.

The implementation is a little messy but I plan to refactor the metaio
classes and this will do for now.