MythGame: Add an in-menu Game Scan.
This is a MythUI game scanner implementation which does a couple of relevant things.

Firstly and most importantly, it allows the user to perform game scans in the "Play Games" UI rather than hunting down the old game scanner in the Setup Menus.  It also makes certain Metadata methods more parallel to the metadata functions in libmythmetadata for simplicity (SaveToDatabase, DeleteFromDatabase, etc. called on the metadata object)

The old scanner is left in place for the short term since this one is a ground-up reimplementation and I may have missed a few things or done something in some wrong way.  For my purposes, though, it's working properly.  I do still have a bit more to do to add a few random features of the old scanner (automatically detecting screenshots, etc).

This commit also does a little more work to extricate the UI code from the database code.  As a result, there's still some repeated functionality here until I'm confident enough to rip out the old stuff in a later refactor.  Still to do is more fully removing the DB code from the UI code, further cleanup of the new scanner, then a reimplementation of the UI code to parallel other MythTV metadata browsers (three view, etc).