Fixes #9972. Apply the construct on first use idiom to two previously staticaly initialized threads.
C++ initializes statics in a strict order within a compilation unit, but across compilation units the order is random.

This is a problem when you statically initialize an MThread because you may initialize it before the static MThread
debugging variables are statically initialized. i.e. mythsocket.o or logging.o may be initialized before mthread.o

It's possible to work around this using compiler attributes, but a cleaner solution is to use the construct-on-first-use idiom as it will work correctly on all systems.

This also adds a warning to the mthread.h doxygen doct to not to statically initialize threads.