Always use the loopback interface when local
If the backend is local to the frontend then we will always use the
loopback interface instead of connecting on the external IP. This adds
some resilience to setups where the network connection may drop, e.g.
using wifi - especially Ubuntu which removes the IP from the interface
if it detects that the network is unreachable.