Remove MySQL DB port ping.
Using telnet to test the MySQL port won't necessarily work, now, since
for MythTV (internal) connections we choose a more-efficient route for
local connections since 50606cd50 .  Because the telnet() function uses
the same MythSocket code that's choosing to use localhost, we weren't
necessarily pinging the MySQL port, so users who configured MySQL to not
listen on localhost and configured their database host using a routable
IP address even when it was run on the same host as their MythTV
applications were unable to run MythTV because it (incorrectly) thought
mysqld wasn't running due to an invalid test's failing.

The primary visible change caused by this patch (other than allowing
users who configure their systems as above to actually run MythTV) is
that the log file for remote backends/frontends will contain a MySQL
error message saying that we were unable to connect to the database when
we're unable to connect to the database, rather than "burying" the
message (by never actually trying the connection and then just trying to
use the UPnP auto-conf as a fallback).  In both cases, we fall back to
UPnP auto-conf, and if unable to find a backend, we exit with the
message, "No UPnP backends found."  Now, however, we'll actually show a
DB connection error, beforehand.

Thanks to Eric Westbrook and Mitch Capper for noticing the issue and to
Mitch Capper for doing some testing for me on IRC, and explaining
exactly how he had configured his system.

Fixes: http://www.gossamer-threads.com/lists/mythtv/dev/489043#489043