Update "Show Program Details" wording.
After 68bd478f changed "Allow this program to re-record" to "Allow this
episode to re-record" in the RecordingPopup to make it consistent with
Previously Recorded's menu item, change "Show Program Details" to "Show
Recording Details" in the same menu (and in the ActionPopup, the only
other location where the same phrase was used).  This seemed more
appropriate, anyway, as the details actually include information about
the recording, itself, such as file size and recording host.

There was also a different translation string, "Show program details"
(different because of differing case) used as the help text for the
DETAILS key binding in tv_play.cpp.  Because this is used for different
details in different screens--program/episode details, recording rule
details, recording details, ...--I changed it to "Show details".

So, after 68bd478f made one less string for translators to translate (as
now there are no instances of "Allow this program to re-record"), this
change forces translators to re-translate "Show Program Details" ->
"Show Recording Details".  It also changes a near-duplicate translation
("Show program details") to be more appropriate in the different
contexts where it's used.