ProgramInfo: add SaveTotalFrames to complement the existing SaveTotalDuration.
The goal of this new field is to allow commflag and the recorder class to store both the total frames and total duration. This will allow the player to display the correct position and duration for programs that are telecined, etc. We will use the following formula to determine the current position.

int position = (frame/total_frames) * duration;

I would also like, if possible, for the recorders to store the wall time of the first frame, current frame and last frame. This will allow the player to query the recorder for the current duration for an in-progress recording. We already have a GetFramesWritten() function to get the total frames. It would then be optimal if the recorder saves the total frames and total duration to programinfo when a recording is completed.