libmythhdhomerun: Update to libhdhomerun 20110810.
The only relevant change here is the addition of a new vchannel status API and struct which allows one to query a vchannel for its protection status.

For purposes of the HDHR Prime, this may indeed be quite useful for performing a "pseudoscan" of the channels from the lineup and only inserting those which can be captured in MythTV.

Since "fetch channels" just calls out to mythfilldatabase, I am thinking about the best way to handle this... perhaps a:

mythfilldatabase --do-channel-updates --pseudoscan --cardid 1

?

For OCUR tuners, this could call the equivalent API for each channel before insertion, in effect performing a periodic channel "scan" that keeps the map entirely up to date.  Probably wouldn't be a good thing to do by default, but users could conceivably add it to their mythfilldatabase args.