Metadata Lookup: conditionally show both Movie and TV results.
When the following are true:

* The item being looked up is a recording/guide item
* The item appears generic or like a movie
* The item has no inetref set

query both the TV and Movie metadata sources and return a list of results for both.  This has a couple of advantages.  The first is that when looking up a generic episode of, say, "Chuck," the old behavior would try the Movie grabber, return some results, and stop.  Instead, in these cases, it's more appropriate to present a list of returns from both results and allow the user to choose.

This also helps with the issue where some clown sets up a metadata record for a TV show at TMDB, and gives the user the opportunity to select the right record from the right metadata source.

All the auto-matching based on year/etc. is still present and functional, and my usual test cases are still automatically matched (Castle, Conan, etc) but with the added bonus of having working auto-match for generic episodes as well.

Nothing for video metadata lookup changes with this commit.  It only affects recordings/MythMetadataLookup.