Fix database logging shutdown
It seems that some of the recent changes have caused issues with the shutdown
of the database logging thread.  The main part of this is that the database
connection was still in scope and deemed active when it was being torn down.
This has been changed.  Also, the original intent (to drain the db queue into
to database on shutdown) has been restored.  Also, the wait at startup was
too long for such short-lived programs as mythpreviewgen, and has been
shortened to not wait after the database is ready.  This may need subsequent
tweaking.

Finally, the call to logStop() in mythcorecontext.cpp has been disabled as it
runs after the call to logStop() in the mythcontext dtor.  There may be some
situations that required this second call, so we will have to keep our eyes
peeled.