Commit [42dfafead] removed a two second sleep in the DB logging startup, unfortunately this can cause DB logging to fail.
The problem is that a MSqlQuery created too soon can cause "Driver not loaded" errors, which don't go away once the MYSQL Qt driver is loaded. This commit recreates the MSqlQuery whenever logqmsg() fails, so that when this happens we will still be able to log to the DB once we manage to get a usable connection to the DB.

This also suppresses printing those error messages at startup & suppresses all DB logging DB errors for one second when it sees a DB error. There isn't much difference between the 1st DBError with logging and the 10,000th one, so this just limits us to logging one such error message per second.