Fix compilation for Windows
Changed the static_cast<uint64_t> to (uint64_t) -- a C-style cast -- as it
works on both Linux and Windows in the intended way.

Also, added PREFIX64 to replace "%llX" with "%I64X" and "%lld" with "%I64d"
on mingw (Windows) builds.  Seems the MS Runtime doesn't do 64-bit in the same
way as glibc.  Figures.  See:
http://msdn.microsoft.com/en-us/library/tcxf1dw6(v=vs.71).aspx

Fixes #9996