Simplify a for loop for the sake of readability
This way it can be understood at a glance, previously it was a bit
harder to understand what was going on.

This would need to be done anyway when we swap out MythDeque
with QList in the future. QList wouldn't return a null pointer if we
called deque on an empty list, it would crash.