Systematically deal with varchar() NOT NULL insertions being done for recordings.
We've been dealing with these as they crop up in error messages since I added the pedandic binding warning to the MSqlQuery. Since these two classes seem to account for most of these errors I decided to look at the schema and do the conversion for all the optional strings with the use of a helper function to keep the code tidy. This means we'll still throw an error if you try to insert a recording without a title, but we won't complain if it doesn't have a category set.