Modify MHEG to use sourceid instead of cardid
This patch modifies the code used to get a channel index to use the
source ID rather than card ID. It simplifies the code a bit but it also
fixes a small problem which only seems to have appeared in the current
git HEAD. When viewing recordings from BBC1, at least, a periodic
message would come up saying "Due to technical difficulties...". This
appears because the BBC MHEG code checks to see which regional variation
it is running on by looking up various transport and service IDs. If the
look-up fails it produces this message. The patch solves this problem.

Signed-off-by: Stuart Morgan <smorgan@mythtv.org>