Revert nanosleep change for Windows.
OK, I shoulda checked first, but my new buildbot slave (not in production yet)
caught it for me.  There is no nanosleep() in mingw, as that is a POSIX only
function.  We do not use cygwin, so we don't get the benefit of all of POSIX.

The list of functions in mingwrt is reputed to be these:
http://msdn.microsoft.com/en-us/library/634ca0c2