Fix gmtime_r in mingw, redo random fix
Seems that the libavformat/utils.h already contains a gmtime_r replacement for
mingw, and it uses an #define.  If we don't check if it's in place, we get
really odd compile errors.  So... if ffmpeg already defined it, and that include
has already been included, no real need to redo it.  If we determine we want
to always use ours, we can just #undef gmtime_r and redefine it.

Also, the qrand() use in Programs.cpp should be fulfilled by just including
compat.h