Fix pxsup2dast compilation
Funny, as soon as you turn on the correct compile flags that were supposed to
be there all the time, this code started to refuse to compile.  To be precise,
there were a pile of forced typecasts using GCC builtins that GCC just would
not use.  They were used in all of one place, which I replaced with a normal
C-style forced typecast that was what they were apparently trying to accomplish.

Also, added a function prototype to keep gcc from warning it was missing, and
removed the obvious error found by cppcheck (double sizeof).