Remove the MBASE_PUBLIC from enum DateTimeFormat
This is causing over 100 warnings on the windows build, and the Linux build
doesn't seem to care one way or the other.  I'm thinking we do not need to
put a public visibility at all onto an enum (there are no functions or data
to affect here).  Anyways, if we do need it, it needs to be done in a different
manner or mingw builds get very whiny.