Take advantage of prepared queries in mythfilldatabase.
Modifies some mythfilldatabase code to prepare queries only once (moved
the prepare calls outside of loops).  Since Qt4 now has true prepared
queries and we now properly re-prepare statements after a reconnect, we
should take advantage of the performance benefit of "reusing" the
already prepared query and just updating the data between executions.
(With old Qt, there was no difference between the 2 approaches, so we
often did the prepare calls inside loops.)

This should provide an observable improvement for everyone's runs (even
if not noticeable).  On my system, marking episode first/last showings
took approximately 3s each before the change and 1.5s each after, with
only 8406 showings (which is a small number compared to the number
expected for a MythTV user with a typical cable or satellite lineup).

And, while I was in there, I modified a log message to be a bit more
formal since it shows at default verbosity.