Increase the maximum linux firewire buffer size
Some JMicron firewire cards will enter into an unrecoverable state if
they run out of buffer space when receiving data. This can cause the
firewire-ohci module to deadlock, leaving a reboot as the only method of
recovery.

Since the consequences are so severe, allow setting the maximum linux
firewire buffer up to 20,000 packets (~3.7 MiB). This should be enough
to buffer up to a second of most HD streams.

Note that the actual buffer size is determined from the HDRingbufferSize
setting. Increasing the value will increase the linux firewire buffer
size linearly.