Extend the character encoding check.
Extend the character-encoding check from 0fd398714 to better approximate
the approach that Qt uses internally to determine the encoding.  Qt has
a much more complex process than is reflected, here, but the approach
below pretty much lines up with "standard" *nix programming and if users
follow the instructions, Qt should use the encoding they specify.  This
also changes the check to specifically look for the string UTF-8
somewhere in the checked environment variables and warns if it's not
there.

This may actually warn in cases where Qt still manages to use a UTF-8
encoding, but since we can't directly find out which encoding Qt is
using and since it's better for users to explicitly request a UTF-8
encoding than rely on some strange fallback that gets them there, this
seems reasonable.